+++ josiah/GibblerAnnealedExpectedLossTrainer.h
@@ -36,7 +36,7 @@ namespace Josiah {
     DerivationCollector m_derivationCollector;
     
     //cache the distribution during gradient calculation
-    std::map<const Derivation*,float> m_p;
+    std::map<const Derivation*,double> m_p;
   };
   
 }
